Add figure to diagram only if it exists
It was possible that the `figure` component of the schema didn't exist
when adding it as a child of the schema diagram.  This caused an error
in `XML::LibXML`:

```
blessed SV reference at /home/cochrane/Projekte/OSSProjekte/MySQL-Workbench-Merge/lib/MySQL/Workbench/Merge.pm line 79.
```

By only adding the `figure` component if it exists causes the first
failing `02_merge` test to pass.

A [post on Perl Monks](https://www.perlmonks.org/?node_id=657872) gave
me a hint as to what the problem could be and hence lead to this
solution.